♻️ refactor(rules): split evaluator.rs into focused modules (#92)
## Why (User Story)
As a developer, I want rule evaluation logic organized by concern so that
template-based detection, composition trees, and filters are independently
maintainable.

## What Changed
- Split 855-line evaluator.rs into 3-file module:
  - mod.rs: RuleEvaluator, SignalScores, signal type mapping
  - composition.rs: Boolean expression tree (AND/OR/NOT) evaluation
  - filters.rs: Post-detection filter application
- Preserved all public API exports

Related to #97
